app/requirements.txt
# PostgreSQL 18 Schema for DEV.to Analytics
# SQLAlchemy Core requirements

# Database
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.9
pgvector>=0.2.4

# Configuration
python-dotenv>=1.0.0

# Date/Time utilities
python-dateutil>=2.8.2

# Optional: Migration tools
alembic>=1.13.0

# Optional: Connection pooling
psycopg2-pool>=1.1

# Development
pytest>=7.4.0
pytest-cov>=4.1.0
